Extracted MIDI note handling from ep2_syn.rb into a standalone class.
